Fixing bug in inventory service: item should be saved in both mongo and postgresql

+++ MicroservicesVersion/ItTrAp.InventoryService/Infrastructure/Persistence/Repositories/MongoMovableItemRepository.cs
@@ -2,21 +2,34 @@ using ItTrAp.InventoryService.Infrastructure.Interfaces;
 using ItTrAp.InventoryService.Infrastructure.Interfaces.Repositories;
 using ItTrAp.InventoryService.Domain.Models;
 using MongoDB.Driver;
+using Microsoft.EntityFrameworkCore;
 
 namespace ItTrAp.InventoryService.Infrastructure.Persistence.Repositories;
 
 public class MongoMovableItemRepository : IRepository<MovableItem, Guid>, IMovableItemRepository
 {
     private readonly IMongoCollection<MovableItem> _collection;
+    private readonly AppDbContext _dbContext;
 
-    public MongoMovableItemRepository(IMongoDatabase database)
+    public MongoMovableItemRepository(IMongoDatabase database, AppDbContext dbContext)
     {
         _collection = database.GetCollection<MovableItem>("movable_items");
+        _dbContext = dbContext;
     }
 
     public async Task<MovableItem> CreateAsync(MovableItem entity, CancellationToken ct = default)
     {
-        await _collection.InsertOneAsync(entity, cancellationToken: ct);
+        try
+        {
+            var savedEntity = await _dbContext.MovableItems.AddAsync(entity, ct);
+            await _dbContext.SaveChangesAsync(ct);
+            await _collection.InsertOneAsync(savedEntity.Entity, cancellationToken: ct);
+        }
+        catch (Exception)
+        {
+            await DeleteAsync(entity.Id, ct);
+            throw;
+        }
         return entity;
     }
 
@@ -37,7 +50,15 @@ public class MongoMovableItemRepository : IRepository<MovableItem, Guid>, IMovab
 
     public async Task<bool> DeleteAsync(Guid id, CancellationToken ct = default)
     {
+        var item = await _dbContext.MovableItems.FirstOrDefaultAsync(i => i.Id == id);
+        if (item == null)
+            return false;
+
         var result = await _collection.DeleteOneAsync(item => item.Id == id, ct);
+
+        _dbContext.MovableItems.Remove(item);
+        await _dbContext.SaveChangesAsync(ct);
+
         return result.DeletedCount > 0;
     }
 }
